refactoring gameinprogresscontent
